requirements.txt
ultralytics
torch
torchvision
opencv-python
numpy
azure-eventhub
python-dotenv
